Added sanitize function
When a string with uppercase and lowercase chars are passed to itemCounts function it does differentiate both versions of the same letter. For example:

itemCounts(['A', 'a', 'B', 'b') would return { a: 1, A: 1, b: 1, B:1 }

For the cases where you need to group both versions of the same letter the sanitize() function can be used. For example:

sanitize('AaBb') would return 'aabb'
then itemCounts(['a', 'a', 'b', 'b']) would return { a: 2, b: 2 }